move dynamic labels filter under the labels filter module
